feat: start work on secured avatar display, avatar images are now saved as webp to save space and bandwith refactor: change default indent for client files to 2 spaces, move user related action from service to controller

+++ client/src/components/modals/confirmation-modal.tsx
@@ -11,39 +11,39 @@ import {
 import { channelService } from "@/services/channelService";
 
 interface ConfirmationDialogProps {
-    channelId: string;
-    isOpen: boolean;
-    onClose: () => void;
+  channelId: string;
+  isOpen: boolean;
+  onClose: () => void;
 }
 
 export function ConfirmationDialog({ channelId, isOpen, onClose }: ConfirmationDialogProps) {
 
-    const handleDelete = async () => {
-        try {
-            await await channelService.deleteChannel(channelId);
-        } catch (err) {
-            console.error("API error deleting channel:", err);
-        }
-        finally {
-            onClose();
-        }
+  const handleDelete = async () => {
+    try {
+      await await channelService.deleteChannel(channelId);
+    } catch (err) {
+      console.error("API error deleting channel:", err);
     }
+    finally {
+      onClose();
+    }
+  }
 
-    return (
-        <AlertDialog open={isOpen} onOpenChange={(open: boolean) => !open && onClose()}>
-            <AlertDialogContent>
-            <AlertDialogHeader>
-                <AlertDialogTitle>Jesteś absolutnie pewny?</AlertDialogTitle>
-                <AlertDialogDescription>
-                Tej akcji nie można cofnąć. Wszystkie wysłane na tym kanale wiadomości oraz załączniki
-                zostaną permanentnie usunięte z serwera.
-                </AlertDialogDescription>
-            </AlertDialogHeader>
-            <AlertDialogFooter>
-                <AlertDialogCancel onClick={() => onClose()}>Anuluj</AlertDialogCancel>
-                <AlertDialogAction onClick={() => handleDelete()}>Kontynuuj</AlertDialogAction>
-            </AlertDialogFooter>
-            </AlertDialogContent>
-        </AlertDialog>
-    )
+  return (
+    <AlertDialog open={isOpen} onOpenChange={(open: boolean) => !open && onClose()}>
+      <AlertDialogContent>
+        <AlertDialogHeader>
+          <AlertDialogTitle>Jesteś absolutnie pewny?</AlertDialogTitle>
+          <AlertDialogDescription>
+            Tej akcji nie można cofnąć. Wszystkie wysłane na tym kanale wiadomości oraz załączniki
+            zostaną permanentnie usunięte z serwera.
+          </AlertDialogDescription>
+        </AlertDialogHeader>
+        <AlertDialogFooter>
+          <AlertDialogCancel onClick={() => onClose()}>Anuluj</AlertDialogCancel>
+          <AlertDialogAction onClick={() => handleDelete()}>Kontynuuj</AlertDialogAction>
+        </AlertDialogFooter>
+      </AlertDialogContent>
+    </AlertDialog>
+  )
 }
